requirements.txt
pickle==4.0
faiss-cpu==1.7.4
find-job-titles==0.7.0
nltk==3.8.1
numpy==1.23.5
pandas==1.5.2
PyPDF2==3.0.1
sentence-transformers==2.2.2
streamlit==1.24.1
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-embedcode==0.1.2
streamlit-extras==0.2.7
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2